refactor: NFT seed処理をdomains.ts内に統合
- createNfts()関数をdomains.ts内に直接定義
- NFT用のヘルパー関数をdomains.tsの上部に配置
- nft.tsファイルを削除してコードを統合
- placeやopportunityと同じパターンに統一

**変更内容:**
- domains.tsにNFT用ヘルパー関数を追加（6個）
- createNfts()関数を追加（createMany + transactionを使用）
- nft.tsモジュールを削除
- インポート文からseedNftsの参照を削除

**一貫性の向上:**
- 全てのseed関数がdomains.ts内に定義される統一されたパターン
- createPlaces、createOpportunitiesと同じ命名規則
